requirements.txt
absl-py==0.12.0
anyio==3.3.0
apache-beam==2.32.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
avro-python3==1.9.2.1
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
crcmod==1.7
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.1.1
Django==3.2.7
dm-tree==0.1.6
docopt==0.6.2
entrypoints==0.3
fastavro==1.4.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.2
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.1
importlib-resources==5.2.2
imutils==0.5.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
jupyterlab-widgets==1.0.1
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
object-detection @ file:///D:/Code/Tima_Onbroading/model_detection/Tensorflow/models/research
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
orjson==3.6.1
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
portalocker==2.3.2
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
Shapely==1.7.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-model-optimization==0.6.0
tensorflow-text==2.6.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
text-unidecode==1.3
tf-models-official==2.6.0
tf-slim==1.1.0
threadpoolctl==2.2.0
tifffile==2020.9.3
tornado==6.1
tqdm==4.62.2
traitlets==4.3.3
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0
